+++ packages/ui/src/index.ts
@@ -11,6 +11,7 @@ export * from './form';
 export * from './icon';
 export * from './icon-button';
 export * from './link';
+export * from './link-button';
 export * as Notice from './notice';
 export * as Popover from './popover';
 export * from './stack';